+++ sam/models/__init__.py
@@ -1,18 +1,36 @@
-from .benchmark import benchmark_wrapper  # noqa: F401
-from .benchmark import (  # noqa: F401
+from .benchmark import benchmark_wrapper
+from .benchmark import (
     benchmark_model,
     plot_score_dicts,
     preprocess_data_for_benchmarking,
 )
-from .keras_templates import create_keras_autoencoder_rnn  # noqa: F401
-from .keras_templates import (  # noqa: F401
+from .keras_templates import create_keras_autoencoder_rnn
+from .keras_templates import (
     create_keras_autoencoder_mlp,
     create_keras_quantile_mlp,
     create_keras_quantile_rnn,
 )
 
-from .sam_shap_explainer import SamShapExplainer  # noqa: F401
-from .linear_model import LinearQuantileRegression  # noqa: F401
-from .base_model import BaseTimeseriesRegressor  # noqa: F401
-from .constant_model import ConstantTimeseriesRegressor  # noqa: F401
-from .mlp_model import MLPTimeseriesRegressor  # noqa: F401
+from .sam_shap_explainer import SamShapExplainer
+from .linear_model import LinearQuantileRegression
+from .base_model import BaseTimeseriesRegressor
+from .constant_model import ConstantTimeseriesRegressor
+from .lasso_model import LassoTimeseriesRegressor
+from .mlp_model import MLPTimeseriesRegressor
+
+__all__ = [
+    "benchmark_wrapper",
+    "benchmark_model",
+    "plot_score_dicts",
+    "preprocess_data_for_benchmarking",
+    "create_keras_autoencoder_rnn",
+    "create_keras_autoencoder_mlp",
+    "create_keras_quantile_mlp",
+    "create_keras_quantile_rnn",
+    "SamShapExplainer",
+    "LinearQuantileRegression",
+    "BaseTimeseriesRegressor",
+    "ConstantTimeseriesRegressor",
+    "LassoTimeseriesRegressor",
+    "MLPTimeseriesRegressor",
+]